Do not spin wait
While spin waiting will produce the most accurate syncing, it's also very CPU intensive (briefly peaking a single core of an i3-330M), and most of the time it's not really needed anyway. Using a timeout of 1ms is better.

Ideally there will be a signaling mechanism between the 2 stream, instead of each stream polling eachother, but I'm keeping this simple.